+++ src/apps/main/bug-report/ErrorDetails.ts
@@ -1,0 +1,30 @@
+/**
+ * Only for error reporting purposes, should not be used
+ * to adjust UI to specific errors for example.
+ * That's what SyncError and SyncFatalError classes are for
+ */
+export type ErrorDetails = {
+  /* Describes in natural language what was being
+   done when this error was thrown */
+  action: string;
+
+  // Message of the original error instance
+  message: string;
+  // Error code of the original error instance
+  code: string;
+  // Stack of the original error instance
+  stack: string;
+
+  /* SYSTEM ERROR SPECIFICS */
+
+  // Error number
+  errno?: number;
+  // System call name
+  syscall?: string;
+  // Extra details about the error
+  info?: Record<string, any>;
+
+  // Aditional info that could be helpful
+  // to debug
+  additionalInfo?: string;
+};
