+++ tests/fixtures/errors.ts
@@ -1,4 +1,5 @@
-import { ErrorCause, AppIssue } from '../../src/apps/shared/types';
+import { AppIssue } from '../../src/apps/shared/types';
+import { ErrorCause } from '../../src/context/virtual-drive/shared/domain/ErrorCause';
 import { AppErrorName } from '../../src/shared/issues/AppIssue';
 import { FatalError } from '../../src/shared/issues/FatalError';
 import { VirtualDriveIssue } from '../../src/shared/issues/VirtualDriveIssue';
@@ -9,12 +10,10 @@ export const createBackupFatalError = (errorName: FatalError) => ({
   errorName,
 });
 
-export const createSyncError = (
-  processErrorName: ErrorCause
-): VirtualDriveIssue => ({
+export const createSyncError = (cause: ErrorCause): VirtualDriveIssue => ({
   name: 'name',
   error: 'DELETE_ERROR',
-  cause: processErrorName,
+  cause: cause,
 });
 
 export const createGeneralIssueFixture = (name: AppErrorName): AppIssue => ({
